+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/internal/utils/Constants.java
@@ -1,6 +1,8 @@
 package com.mapbox.navigation.base.internal.utils;
 
-public class Constants {
+public final class Constants {
 
-    public static final String KEY_WAYPOINTS = "waypoints";
+    public static final class RouteResponse {
+        public static final String KEY_WAYPOINTS = "waypoints";
+    }
 }